move some auth-related code into src/features/auth
