+++ src/features/README.md
@@ -1,0 +1,3 @@
+The features folder is for large chunks of app-specific functionality, such as
+authentication. Group files not by file type, but by interdependencies & high
+likelihood of needing to be changed together.
