+++ src/features/auth/useLogout.ts
@@ -1,0 +1,12 @@
+import { useRecoilLoadCatch } from 'hooks';
+
+import { clearStateAfterLogout } from './useFinishAuth';
+
+export const useLogout = () => {
+  return useRecoilLoadCatch(
+    ({ set }) =>
+      async () =>
+        clearStateAfterLogout(set),
+    []
+  );
+};
